Keep `body_to_bytes` in cassette.py rather than a module of its own
`_shared` already imports from `cassette`, so the helper can sit next to
its caller and ride that edge instead of needing a third module to break
the cycle.
