+++ src/cassetter/cassette.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import fnmatch
+import json
 import os
 import re
 import stat
@@ -30,7 +31,6 @@ from cassetter._core import (
     scrub_interaction,
     scrub_ws_interaction,
 )
-from cassetter._wire import body_to_bytes
 from cassetter.intercept._base import is_localhost
 from cassetter.introspection import RecordedRequest, recorded_request
 from cassetter.recording import RecordMode
@@ -545,6 +545,27 @@ class Cassette:
 
         self._inner.add_ws_interaction(interaction)
         self._dirty = True
+
+
+def body_to_bytes(body: Body) -> bytes:
+    """Serialize a recorded body to the wire bytes of a replayed response.
+
+    A JSON body is stored parsed, so the original bytes are gone by the time
+    this runs: the output is stdlib `json.dumps` formatting, not a byte-for-byte
+    reproduction of what the server sent. Callers that recompute Content-Length
+    must do so from these bytes.
+
+    `body.content` materializes a fresh Python object on every read, so each
+    branch reads it once.
+    """
+    content = body.content
+    if body.body_type == "json":
+        return json.dumps(content).encode()
+    if body.body_type == "text":
+        return content.encode() if isinstance(content, str) else b""
+    if body.body_type == "binary":
+        return content if isinstance(content, bytes) else b""
+    return b""
 
 
 def _get_header(headers: dict[str, list[str]], name: str) -> str | None:
