+++ src/cassetter/intercept/_shared.py
@@ -1,7 +1,11 @@
 from __future__ import annotations
 
-from cassetter._wire import body_to_bytes as body_to_bytes
-from cassetter.cassette import BeforeRecordRequest, RawRequest, SkipRecording
+from cassetter.cassette import (
+    BeforeRecordRequest,
+    RawRequest,
+    SkipRecording,
+    body_to_bytes as body_to_bytes,
+)
 
 
 def apply_before_record_request(
